requirements.txt
boto3
botocore
requests
bandit
awsiotsdk

